Soundboard/requirements.txt
sounddevice
soundfile
pydub
keyboard
tkinterdnd2
Pillow
json5
pygame
PySide6
